Issue #960 remove require cycle by splitting up header component code into new file
